refactor: 함수 분리 및 불필요한 출력 삭제
