+++ crawling_restaurnats_detail.py
@@ -1,16 +1,17 @@
-import time
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
-import traceback
 import re
+
+from crawling_menu import calculate_menu_average, crawling_menu
+
+except_keywords = ["카페", "디저트", "베이커리"]
 
 def crawling_restaurant_info(driver, restaurant_elements):
     restaurant_info = []
-    except_keywords = ["카페", "디저트", "베이커리"]
     print(f"[{len(restaurant_elements)}개의 가게 디테일 정보 크롤링 시작]")
     
-    for idx, restaurant in enumerate(restaurant_elements, start=0):
+    for idx, restaurant in enumerate(restaurant_elements, start = 0):
         try:
             store_name = restaurant.find_element(By.CSS_SELECTOR, ".TYaxT").text.strip()
             category = restaurant.find_element(By.CSS_SELECTOR, ".KCMnt").text.strip()
@@ -24,20 +25,17 @@ def crawling_restaurant_info(driver, restaurant_elements):
 
             driver.switch_to.default_content()
             WebDriverWait(driver, 60).until(EC.presence_of_element_located((By.XPATH, '//*[@id="entryIframe"]')))
-
-            iframe = driver.find_element(By.XPATH, '//*[@id="entryIframe"]')
-            iframe_src = iframe.get_attribute("src")
-            driver.switch_to.frame(iframe)
-
-            place_id = re.search(r"/place/(\d+)", iframe_src).group(1)
+            
+            entryIframe = driver.find_element(By.XPATH, '//*[@id="entryIframe"]')
+            entryIframe_src = entryIframe.get_attribute("src")
+            place_id = re.search(r"/place/(\d+)", entryIframe_src).group(1)
+        
+            driver.switch_to.frame(entryIframe)
+        
             print(f"{idx} 가게명 : ", store_name)
-            print("[주소 정보 크롤링 시작]")
             address = crawling_address(driver)
-            print("[이미지 정보 크롤링 시작]")
             images = crawling_img(driver)
-            print("[영업 시간 정보 크롤링 시작]")
             opening_hours = crawling_openingHours(driver)
-            print("[메뉴 정보 크롤링 시작]")
             menu = crawling_menu(driver)
 
             restaurant_info.append({
@@ -51,18 +49,19 @@ def crawling_restaurant_info(driver, restaurant_elements):
                 "menu_average" : int(calculate_menu_average(menu))
             })
 
-            driver.switch_to.default_content()
-            WebDriverWait(driver, 60).until(EC.frame_to_be_available_and_switch_to_it((By.ID, "searchIframe")))
-            del restaurant
-            
+            switch_iframe(driver, "searchIframe")
         except Exception :
-            driver.switch_to.default_content()
-            WebDriverWait(driver, 60).until(EC.frame_to_be_available_and_switch_to_it((By.ID, "searchIframe")))
+            switch_iframe(driver, "searchIframe")
             continue
             
     return restaurant_info
 
+def switch_iframe(driver, name):
+    driver.switch_to.default_content()
+    WebDriverWait(driver, 60).until(EC.frame_to_be_available_and_switch_to_it((By.ID, name)))
+
 def crawling_address(driver):
+    print("[주소 정보 크롤링 시작]")
     try:
         address_elements = WebDriverWait(driver, 60).until(EC.presence_of_element_located((By.CSS_SELECTOR, ".place_section_content .vV_z_")))
         address = address_elements.find_element(By.CSS_SELECTOR, "span.LDgIH").text.strip()
@@ -71,6 +70,7 @@ def crawling_address(driver):
     return address
 
 def crawling_img(driver):
+    print("[이미지 정보 크롤링 시작]")
     try:
         images = []
         image_elements = WebDriverWait(driver, 60).until(EC.visibility_of_element_located((By.CSS_SELECTOR, ".CB8aP .uDR4i")))
@@ -81,9 +81,9 @@ def crawling_img(driver):
     return images
 
 def crawling_openingHours(driver):
+    print("[영업 시간 정보 크롤링 시작]")
     try:
         if driver.find_elements(By.CSS_SELECTOR, ".O8qbU.J1zN9 span.LDgIH"):
-            print("영업시간 없습니다")
             return []
         
         button_element = WebDriverWait(driver, 60).until(EC.element_to_be_clickable((By.CSS_SELECTOR, ".gKP9i.RMgN0")))
@@ -144,75 +144,3 @@ def parse_hours(hours):
 def clean_time_string(t: str) -> str:
     t = t.strip().replace(",", "")
     return "00:00" if t == "24:00" else t
-
-def calculate_menu_average(menus):
-    main_prices = [menu["price"] for menu in menus
-               if menu.get("isMain") and isinstance(menu.get("price"), int)]
-
-    if not main_prices:
-        main_prices = [menu["price"] for menu in menus
-                if isinstance(menu.get("price"), int)]
-    
-    return sum(main_prices) / len(main_prices)
-
-def crawling_menu(driver):
-    try:
-        menus = []
-        a_elements = driver.find_elements(By.CSS_SELECTOR, ".place_fixed_maintab a")
-        menu_link = next((a for a in a_elements if "menu" in a.get_attribute("href")), None)
-
-        if menu_link is None:
-            return menus
-        
-        driver.execute_script("arguments[0].click();", menu_link)
-        time.sleep(0.1)
-        if driver.find_elements(By.CSS_SELECTOR, ".smart_category.slick-slider.general_place"):
-            print("발견됨 !!")
-            return menus
-        
-        driver.switch_to.default_content()
-        WebDriverWait(driver, 60).until(EC.presence_of_element_located((By.ID, "entryIframe")))
-        driver.switch_to.frame(driver.find_element(By.ID, "entryIframe"))
-
-        menu_elements = WebDriverWait(driver, 60).until(EC.presence_of_all_elements_located((By.CSS_SELECTOR, ".place_section_content .E2jtL")))
-    
-        for m in menu_elements:
-            try:
-                is_main = bool(m.find_elements(By.CSS_SELECTOR, "span.place_blind"))
-            except:
-                is_main = False
-
-            try:
-                name = m.find_element(By.CSS_SELECTOR, "span.lPzHi").text.strip()
-            except:
-                name = None
-
-            try:
-                introduce = m.find_element(By.CSS_SELECTOR, ".kPogF").text.strip()
-            except:
-                introduce = None
-
-            try:
-                price = m.find_element(By.CSS_SELECTOR, ".GXS1X em").text.strip()
-                price = int(price.replace(",", ""))
-            except:
-                price = None
-
-            try:
-                img_el = m.find_element(By.CSS_SELECTOR, ".place_thumb img")
-                img_src = img_el.get_attribute("src").strip()
-            except:
-                img_src = None
-
-            menus.append({
-                "isMain": is_main,
-                "name": name,
-                "introduce": introduce,
-                "price": price,
-                "imgUrl": img_src
-            })
-            
-    except Exception as e:
-         print(" 메뉴 에러 발생 : " + str(e))
-         traceback.print_exc()
-    return menus
